nola-god-level/requirements.txt
Flask==3.0.3
streamlit==1.39.0
pandas==2.2.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
matplotlib==3.9.2
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
Faker==20.1.0